Transform result models to CSV
Since using the tables for CSV generation is unreliable
and disruptive (as the user may be working on one scenario,
which could trigger an export, thus cycling through all
scenarios and disrupting the user experience), we switch to
using the models for exporting.

The results are identical to the ones previously being
generated.